feat: use protobuf field names in error messages
Use the name of the field from the protobuf schema instead of the
generated golang implemenatation. Make it easier to understand generated
validation errors returned to clients. A client program are written
in languages other than Golang. It is easy for developers working
with other languages to look at the protobuf field name and know what
the generated code looks like for their language. It is more work to
require all engineers understand the Golang code generation naming
rules.